fix types to satisfy -Wextra
There is lot of work to be done to satisfy -Wextra.
This is the first step where I have eliminated all
errors related to the non-strict type coercion.

Reduced the scope of some variables.

While at it I have eliminated function one_cmd as making
code more complex that it has to be.

Updates #443

+++ dat.h
@@ -274,9 +274,9 @@ struct Conn {
     int    pending_timeout;
     char   halfclosed;
 
-    char cmd[LINE_BUF_SIZE]; // this string is NOT NUL-terminated
-    int  cmd_len;
-    int  cmd_read;
+    char   cmd[LINE_BUF_SIZE]; // this string is NOT NUL-terminated
+    size_t cmd_len;
+    int    cmd_read;
 
     char *reply;
     int  reply_len;
